refactor(backend): split ai route helpers
